requirements.txt
openai==0.27.0
python-decouple==3.8